refactor: factorize FederationGroupMapper and FederationRoleMapper classes
