+++ lib/ansible/modules/identity/keycloak/keycloak_ldap_role_mapper.py
@@ -5,13 +5,6 @@
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 from __future__ import absolute_import, division, print_function
 
-from ansible.module_utils.identity.keycloak.urls import COMPONENTS_BY_UUID_URL, MAPPER_BY_NAME
-from ansible.module_utils.identity.keycloak.utils import (
-    clean_payload_with_config,
-    snake_to_point_case,
-    if_absent_add_a_default_value,
-)
-
 __metaclass__ = type
 
 ANSIBLE_METADATA = {'metadata_version': '1.1', 'status': ['preview'], 'supported_by': 'community'}
@@ -325,120 +318,28 @@ group_mapper:
           sample: admin-cli 
 '''
 
-from copy import deepcopy
-
 from ansible.module_utils.identity.keycloak.keycloak import (
     KeycloakError,
-    get_on_url,
     keycloak_argument_spec,
     get_token,
-    put_on_url,
-    delete_on_url,
-    post_on_url,
 )
-from ansible.module_utils.identity.keycloak.keycloak_ldap_federation import (
-    LdapFederationBase,
-    COMPONENTS_URL,
-)
-from ansible.module_utils.six.moves.urllib.parse import quote
-from ansible.module_utils.common.dict_transformations import recursive_diff, dict_merge
+from ansible.module_utils.common.dict_transformations import dict_merge
 
 from ansible.module_utils.identity.keycloak.crud import crud_with_instance
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.identity.keycloak.keycloak import KeycloakAPI
+from ansible.module_utils.identity.keycloak.keycloak_ldap_mapper import FederationMapper
+from ansible.module_utils.identity.keycloak.utils import (
+    snake_to_point_case,
+    if_absent_add_a_default_value,
+)
 
 
-class FederationRoleMapper(object):
+class FederationRoleMapper(FederationMapper):
     def __init__(self, module, connection_header):
-        self.module = module
-        self.restheaders = connection_header
-        self.uuid = self.module.params.get('mapper_uuid')
-        self.description = 'role mapper {}'.format(self.given_id)
-        if self.module.params.get('mapper_name'):
-            self.federation = LdapFederationBase(module, connection_header)
-            if not self.federation.uuid:
-                raise KeycloakError(
-                    'Cannot access mapper because {given_id} federation does not exist.'.format(
-                        given_id=self.federation.given_id
-                    )
-                )
-        else:
-            self.federation = None
-        self.initial_representation = self.representation
-        try:
-            self.uuid = self.initial_representation['id']
-        except KeyError:
-            pass
-        else:
-            if self.initial_representation['providerId'] != 'role-ldap-mapper':
-                raise KeycloakError(
-                    '{given_id} is not a role mapper.'.format(given_id=self.given_id)
-                )
-
-    @property
-    def given_id(self):
-        """Get the asked mapper id given by the user.
-
-        :return the asked id given by the user as a name or an uuid.
-        :rtype: str
-        """
-        if self.module.params.get('mapper_name'):
-            return self.module.params.get('mapper_name')
-        return self.module.params.get('mapper_uuid')
-
-    @property
-    def representation(self):
-        return clean_payload_with_config(
-            get_on_url(
-                url=self._get_mapper_url(),
-                restheaders=self.restheaders,
-                module=self.module,
-                description=self.description,
-            )
+        super(FederationRoleMapper, self).__init__(
+            module=module, connection_header=connection_header, mapper_type='role'
         )
-
-    def _get_mapper_url(self):
-        """Create the url in order to get the federation from the given argument (uuid or name)
-        :return: the url as string
-        :rtype: str"""
-        if self.uuid:
-            return COMPONENTS_BY_UUID_URL.format(
-                url=self.module.params.get('auth_keycloak_url'),
-                realm=quote(self.module.params.get('realm')),
-                uuid=quote(self.uuid),
-            )
-        return MAPPER_BY_NAME.format(
-            url=self.module.params.get('auth_keycloak_url'),
-            realm=quote(self.module.params.get('realm')),
-            mapper_name=quote(self.module.params.get('mapper_name').lower()),
-            federation_uuid=self.federation.uuid,
-        )
-
-    def _arguments_update_representation(self):
-        clean_payload = clean_payload_with_config(self._create_payload(), credential_clean=False)
-        payload_without_empty_values = deepcopy(clean_payload)
-        for key, value in clean_payload['config'].items():
-            if not value:
-                payload_without_empty_values['config'].pop(key)
-        payload_diff, _ = recursive_diff(payload_without_empty_values, self.initial_representation)
-        try:
-            config_diff = payload_diff.pop('config')
-        except KeyError:
-            config_diff = {}
-        if not payload_diff and not config_diff:
-            return False
-        return True
-
-    def update(self, check=False):
-        if not self._arguments_update_representation():
-            return {}
-        payload = self._create_payload()
-        if check:
-            return clean_payload_with_config(payload)
-        put_on_url(
-            self._get_mapper_url(), self.restheaders, self.module, self.description, payload
-        )
-        return clean_payload_with_config(payload)
 
     def _create_payload(self):
         translation = {'mapper_name': 'name', 'mapper_uuid': 'id'}
@@ -513,27 +414,10 @@ class FederationRoleMapper(object):
                     )
                 )
 
-    def delete(self):
-        delete_on_url(self._get_mapper_url(), self.restheaders, self.module, self.description)
-
     def create(self, check=False):
         if not self.module.params.get('roles_dn'):
             raise KeycloakError('roles_dn is mandatory for role mapper creation.')
-        payload = self._create_payload()
-        if check:
-            return clean_payload_with_config(payload)
-        post_url = COMPONENTS_URL.format(
-            url=self.module.params.get('auth_keycloak_url'),
-            realm=quote(self.module.params.get('realm')),
-        )
-        post_on_url(
-            url=post_url,
-            restheaders=self.restheaders,
-            module=self.module,
-            description=self.description,
-            representation=payload,
-        )
-        return clean_payload_with_config(payload)
+        super(FederationRoleMapper, self).create(check)
 
 
 def run_module():
@@ -562,10 +446,9 @@ def run_module():
         ),
         memberof_ldap_attribute=dict(
             type=str
-        ),  # used only if user roles retrieve strategy is GET_ROLES_FROM_USER_MEMBEROF_ATTRIBUTE
+        ),
         use_realm_roles_mapping=dict(type=bool),
-        # the value should be in the list of client id of the realm (to check)
-        client_id=dict(type=str),  # used only if user realm roles mapping is False.
+        client_id=dict(type=str),
     )
     argument_spec = keycloak_argument_spec()
     argument_spec.update(meta_args)
